Jar File
Added Build from 1/16/2016 @ 3:13pm
Removed some redundant and useless code.

+++ core/src/com/icnhdevelopment/wotn/world/World.java
@@ -306,7 +306,7 @@ public class World {
         if (battleStage>-1){
             batch.setProjectionMatrix(inventory.getRenderCam().combined);
             batch.begin();
-            TextureRegion tr = new TextureRegion(battleTransition, (int)((battleStage%4)*160), (int)Math.floor((battleStage/4)*90), 160, 90);
+            TextureRegion tr = new TextureRegion(battleTransition, ((battleStage%4)*160), (int)Math.floor((battleStage/4)*90), 160, 90);
             batch.draw(tr, 0, 0, Game.WIDTH(), Game.HEIGHT());
             batch.end();
             batch.setProjectionMatrix(camera.combined);
